% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfbydate.R
\name{dfbydate}
\alias{dfbydate}
\title{IPOS: Key information of patent applications by lodgement date}
\usage{
dfbydate(date)
}
\arguments{
\item{date}{The particular lodgement date, in the following format and in inverted commas: \code{"YYYY-MM-DD"}.}
}
\value{
Dataframe of key information of patent applications on the particular lodgement date.
}
\description{
This function is for the API of Intellectual Property Office of Singapore (IPOS). It retrieves key information* on patent applications by lodgement date and tidies it into a clean dataframe. More info on API here: \url{https://data.gov.sg/dataset/ipos-apis?resource_id=6a030bf2-22da-4621-8ab0-9a5956a30ef3}.
}
\details{
This function does not use the \code{jsonlite} package to parse the API's JSON  output (function was built before discovering \code{jsonlite} and building function \code{getbydate} in this package). Instead it cleans the data from the JSON list of lists into an easily understandable dataframe. *Key information include the following variables from the API: \code{"summary.applicationNum"}, \code{"summary.applicationStatus"}, \code{"summary.applicationType"}, \code{"summary.filingDate"}, \code{"summary.lodgementDate"}, \code{"summary.titleOfInvention"}, \code{"inventors.name"}, \code{"inventors.address"}, \code{"inventors.countryOfResidence.description"}, \code{"inventors.countryOfResidence.code"}, \code{"inventors.nationality"}, \code{"applicant.uenCompanyCode"}, \code{"applicant.name"}, \code{"applicant.address"}, \code{"applicant.countryOfIncorporationOrResidence.description"}, \code{"applicant.stateOfIncorporation.description"}, \code{"applicant.stateOfIncorporation.code"}, \code{"applicant.nationality.description"}.
}
