% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dftwodates.R
\name{dftwodates}
\alias{dftwodates}
\title{IPOS: Key information of patent applications lodged between two dates (both dates inclusive)}
\usage{
dftwodates(startdate, enddate = Sys.Date())
}
\arguments{
\item{startdate}{The start date of lodgement date period, in the following format and in inverted commas: \code{"YYYY-MM-DD"}.}

\item{enddate}{The end date of lodgement date period, in the following format and in inverted commas: \code{"YYYY-MM-DD"}.}
}
\value{
Dataframe of key information of patent applications lodged between two dates (both dates inclusive).
}
\description{
This function is for the API of Intellectual Property Office of Singapore (IPOS). It retrieves key information* on patent applications between two lodgement dates (both dates inclusive) and tidies it into a clean dataframe. More info on API here: \url{https://data.gov.sg/dataset/ipos-apis?resource_id=6a030bf2-22da-4621-8ab0-9a5956a30ef3}.
}
\details{
This function is a cumulative version of the \code{dfbydate} function in this package. The API Client is only able to retrieve applications by one date, and often analyses is done across multiple dates in one period. *Key information include the following variables from the API: \code{"summary.applicationNum"}, \code{"summary.applicationStatus"}, \code{"summary.applicationType"}, \code{"summary.filingDate"}, \code{"summary.lodgementDate"}, \code{"summary.titleOfInvention"}, \code{"inventors.name"}, \code{"inventors.address"}, \code{"inventors.countryOfResidence.description"}, \code{"inventors.countryOfResidence.code"}, \code{"inventors.nationality"}, \code{"applicant.uenCompanyCode"}, \code{"applicant.name"}, \code{"applicant.address"}, \code{"applicant.countryOfIncorporationOrResidence.description"}, \code{"applicant.stateOfIncorporation.description"}, \code{"applicant.stateOfIncorporation.code"}, \code{"applicant.nationality.description"}.
}
