% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbydate.R
\name{getbydate}
\alias{getbydate}
\title{IPOS: Raw data of applications by lodgement date}
\usage{
getbydate(date, type = "patents")
}
\arguments{
\item{date}{The particular lodgement date, in the following format and in inverted commas: \code{"YYYY-MM-DD"}.}

\item{type}{Three choices for the type of IP applications: \code{patents}, \code{designs} or \code{trademarks}. Inputs in inverted commas e.g. \code{"patents"}.}
}
\value{
Dataframe of raw data from IPOS API (parsed from JSON format).
}
\description{
This function is for the API of Intellectual Property Office of Singapore (IPOS). It retrieves raw data of real-time patent, design or trademark applications from the API, and parses the JSON output into a dataframe. More info on API here: \url{https://data.gov.sg/dataset/ipos-apis?resource_id=6a030bf2-22da-4621-8ab0-9a5956a30ef3}.
}
\details{
Note: The parsing of raw data in JSON format using the \code{jsonlite} package results in some values being unclean or in an unsuitable format. The function \code{dfbydate} in this package provides a dataframe with select key information of applications (patents only) that is cleaned.
}
