% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarpatent.R
\name{similarpatent}
\alias{similarpatent}
\title{IPOS: Similar patent applications lodged between two dates}
\usage{
similarpatent(startdate, enddate = Sys.Date(), phrase)
}
\arguments{
\item{startdate}{The start date of lodgement date period, in the following format and in inverted commas: \code{"YYYY-MM-DD"}.}

\item{enddate}{The end date of lodgement date period, in the following format and in inverted commas: \code{"YYYY-MM-DD"}.}

\item{phrase}{The phrase to search for in patent applications' title of invention. Phrase in inverted commas e.g. \code{"example"}.}
}
\value{
Dataframe of patent applications lodged between two dates (both dates inclusive), with title of invention containing phrase of interest.
}
\description{
This function is for the API of Intellectual Property Office of Singapore (IPOS). It allows you to scan through patent application between two lodgement dates (both dates inclusive) for those with titles of invention that contain your phrase of interest. More info on API here: \url{https://data.gov.sg/dataset/ipos-apis?resource_id=6a030bf2-22da-4621-8ab0-9a5956a30ef3}.
}
